+++ app/jobs/jms/manage_message_job.rb
@@ -5,7 +5,7 @@ require "nori"
 class Jms::ManageMessageJob < ApplicationJob
   queue_as :pc_subscriber
 
-  def perform(message, eic_base_url, logger)
-    Jms::ManageMessage.new(message, eic_base_url, logger)
+  def perform(message, eosc_registry_base_url, logger)
+    Jms::ManageMessage.new(message, eosc_registry_base_url, logger)
   end
 end