+++ lib/jms/subscriber.rb
@@ -14,7 +14,7 @@ module Jms
     end
 
     def initialize(topic, login, pass,  host,
-                   client_name, eic_base_url,
+                   client_name, eosc_registry_base_url,
                    ssl_enabled,
                    token = nil,
                    client: Stomp::Client,
@@ -24,7 +24,7 @@ module Jms
       @client = client.new(conf_hash(login, pass, host, client_name, ssl_enabled))
       log "Parameters: #{conf_hash(login, pass, host, client_name, ssl_enabled)}"
       @destination = topic
-      @eic_base_url = eic_base_url
+      @eosc_registry_base_url = eosc_registry_base_url
       @token = token
     end
 
@@ -32,7 +32,7 @@ module Jms
       log "Start subscriber on destination: #{@destination}"
       @client.subscribe("/topic/#{@destination}.>", { "ack": "client-individual", "activemq.subscriptionName": "mpSubscription" }) do |msg|
         log "Arrived message"
-        Jms::ManageMessage.new(msg, @eic_base_url, @logger, @token).call
+        Jms::ManageMessage.new(msg, @eosc_registry_base_url, @logger, @token).call
         @client.ack(msg)
       rescue Jms::ManageMessage::ResourceParseError, Jms::ManageMessage::WrongMessageError, JSON::ParserError, StandardError => e
         @client.unreceive(msg)