+++ spec/controllers/concerns/searchable_spec.rb
@@ -260,10 +260,10 @@ RSpec.describe ApplicationController, type: :controller do
       let!(:service1) { create(:service) }
       let!(:service2) { create(:service) }
       let!(:service4) { create(:service) }
-      let!(:service_source1) { create(:service_source, source_type: "eic", service_id: service1.id) }
-      let!(:service_source2) { create(:service_source, source_type: "eic", service_id: service2.id) }
+      let!(:service_source1) { create(:service_source, source_type: "eosc_registry", service_id: service1.id) }
+      let!(:service_source2) { create(:service_source, source_type: "eosc_registry", service_id: service2.id) }
       let!(:filter_class) { Filter::UpstreamSource }
-      let!(:values) { "eic" }
+      let!(:values) { "eosc_registry" }
       let!(:param_name) { :source }
       let!(:field_name) { :source }
       it "checks if upstream source filter works" do